pyproject.toml
[tool.poetry]
name = "pickle-bot"
version = "1.0.0"
description = "A Discord bot to handle pickleball-related needs"
authors = ["topol <topologicalinvariant@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
bot = "pickle_bot.bot:main"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
discord-py = "^2.3.2"


[tool.poetry.group.dev.dependencies]
python-lsp-server = "^1.9.0"
flake8 = "^6.1.0"
flake8-black = "^0.3.6"
black = "^23.11.0"
flake8-import-order = "^0.18.2"
flake8-bugbear = "^23.9.16"
flake8-pyproject = "^1.2.3"
pyupgrade = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
select = ["B", "B9", "BLK", "C", "E", "F", "I", "W"]
ignore = ["E203", "E501", "W503"]
max-complexity = 10
max-line-length = 100
application-import-names = ["pickle_bot", "tests"]
import-order-style = "google"

[tool.black]
line-length = 100
preview = true

